method to show overdue task

+++ src/Logic/TNotesLogic.java
@@ -283,7 +283,9 @@ public class TNotesLogic {
 	// if i want to hold the main Array list
 	public TaskFile deleteTask(ArrayList<String> fromParser) throws Exception {
 		fromParser.remove(0);
-		assertNotEquals(0, fromParser.size());
+		if(fromParser.isEmpty()) {
+			throw new Exception("invalid command");
+		}
 		return storage.deleteTask(fromParser.get(0));
 	}
 
@@ -664,7 +666,7 @@ public class TNotesLogic {
 			if (storage.addRecurringTask(recurTask)) {
 				return recurTask;
 			} else {
-				System.out.println("did not manage to add to storage");
+				throw new Exception("did not add to storage");
 			}
 		} else if (type.equals("startTime")) {
 			storage.deleteTask(title);
@@ -673,7 +675,7 @@ public class TNotesLogic {
 			if (storage.addRecurringTask(recurTask)) {
 				return recurTask;
 			} else {
-				System.out.println("did not manage to add to storage");
+				throw new Exception("did not add to storage");
 			}
 		} else if (type.equals("endTime")) {
 			storage.deleteTask(title);
@@ -681,7 +683,7 @@ public class TNotesLogic {
 			if (storage.addRecurringTask(recurTask)) {
 				return recurTask;
 			} else {
-				System.out.println("did not manage to add to storage");
+				throw new Exception("did not add to storage");
 			}
 		} else if (type.equals("startDate")) {
 			storage.deleteTask(title);
@@ -689,7 +691,7 @@ public class TNotesLogic {
 			if (storage.addRecurringTask(recurTask)) {
 				return recurTask;
 			} else {
-				System.out.println("did not manage to add to storage");
+				throw new Exception("did not add to storage");
 			}
 		} else if (type.equals("endDate")) {
 			storage.deleteTask(title);
@@ -697,7 +699,7 @@ public class TNotesLogic {
 			if (storage.addRecurringTask(recurTask)) {
 				return recurTask;
 			} else {
-				System.out.println("did not manage to add to storage");
+				throw new Exception("did not add to storage");
 			}
 		} else if (type.equals("details")) {
 			storage.deleteTask(title);
@@ -705,7 +707,7 @@ public class TNotesLogic {
 			if (storage.addRecurringTask(recurTask)) {
 				return recurTask;
 			} else {
-				System.out.println("did not manage to add to storage");
+				throw new Exception("did not add to storage");
 			}
 		} else if (type.equals("important")) {
 			storage.deleteTask(title);
@@ -718,12 +720,22 @@ public class TNotesLogic {
 				return currentFile;
 			}
 		} else
-			System.out.println("did not edit");
+			throw new Exception("did not edit");
 		return currentFile;
 	}
 
 	public TaskFile deleteRecurringTask(ArrayList<String> fromParser) throws Exception {
+		fromParser.remove(0);
+		if(fromParser.isEmpty()) {
+			throw new Exception("invalid command");
+		}
 		return storage.deleteTask(fromParser.get(0));
 	}
-
+	public ArrayList<TaskFile> callOverdueTasks() throws Exception{
+		ArrayList<TaskFile> listOfOverdueTasks = storage.retrieveOverdueTasks();
+		if(listOfOverdueTasks.isEmpty()) {
+			throw new Exception("no overdue tasks");
+		}
+		return listOfOverdueTasks;
+	}
 }
